ARM: OMAP: AM335x: Add support for Beta GP EVM.
This patch adds support for BETA EVM.
BETA EVM is set as the default configuration.
Since the peripherals were supported only in profile 0 for
alpha EVM, It makes use of the exisiting GP EVM structure.

Changes have been made to MMC and audio.

Flag gp_evm_revision can be used to differentiate between the
revisions of the GP EVM.

Signed-off-by: Patil, Rachna <rachna@ti.com>