usb::musb::cppi41_dma: Check if scheduling is required after DMA
Even after TX DMA, data can still be in the FIFO and we would keep
rescheduling the worker thread in a polling fashion taking up a lot
of CPU if the controller was slow to respond. We check if rescheduling
is required and sleep if so.

This fixes high CPU load issue seen with g_mass_storage on a BeagleBone
and could potentially fix similar issues seen with other gadget drivers.

v2 changes:
Avoid rescheduling a "worker", this process involves a lot of workqueue
related crunching that is unnecessary. If we're polling, we might as well
do it with minimum overhead, and giving other processes CPU time at the
same time.

Signed-off-by: Joel A Fernandes <joelagnel@ti.com>