ARM: OMAP: AM33XX: PM: Get rid of hardcoded resume address
Instead of using a harcoded resume address in the IPC registers
add a dynamically computed offset to the start of OCMC address.

Note: This approach currently assumes that the suspend-resume code
is the first block of code that is copied to internal RAM. Due to
the initcall level used by the SRAM mapping code this holds true.
Ideally the sram mapping code should provide an API for doing this.
This will be done at a later point of time when the generic SRAM
allocator gets merged in mainline kernel.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>