ARM: OMAP: AM33XX: PM: Restore the PLLs to pre-suspend state
In some scenarios all the PLLs (especially Display) might not be
in a locked state when the suspend process starts. Trying to relock
a PLL in MN-bypass mode without a proper set of M and N values results
into the PLLs never locking and the resume process hanging. Fix this
by restoring the PLLs to the mode that they were in before the suspend
process started.

Signed-off-by: Vaibhav Bedia <vaibhav.bedia@ti.com>