capebus: Core capebus support
Introducing capebus; a bus that allows small boards (capes) to connect
to a complex SoC using simple expansion connectors.

Up to now to support these kind of boards, one had to hack the board files,
and do all sort of gymnastics to handle all the different cases of
conflict resolution.

Capebus provides abstractions that keep the pain to a minimum.

This part of the series is introducing the core capebus functionality
dealing with the basic bus & driver probe functions.

Signed-off-by: Pantelis Antoniou <panto@antoniou-consulting.com>